PIM-6353: Move product model normalizer into dedicated folder
