+++ src/Pim/Component/Catalog/Normalizer/Indexing/ProductAndModel/ProductModelNormalizer.php
@@ -1,9 +1,8 @@
 <?php
 
-namespace Pim\Component\Catalog\Normalizer\Indexing;
+namespace Pim\Component\Catalog\Normalizer\Indexing\ProductAndModel;
 
 use Pim\Component\Catalog\Model\ProductInterface;
-use Pim\Component\Catalog\Normalizer\Standard\ProductNormalizer as StandardProductNormalizer;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
